Write the EXIF orientation as a SHORT so rotation takes effect
Rotation rewrote the Orientation tag through the library's byte
overload, which stores the tag with the EXIF type BYTE. Orientation has
to be a SHORT. Readers other than this app reject the tag outright, and
on a file whose EXIF block is big-endian (camera or scanner output) the
app's own reader reads the value in the file's byte order, gets a value
that is not a valid orientation and falls back to "not rotated". A
rotation therefore looked like it had done nothing, and repeating it did
not accumulate either.

Files that have no EXIF block were unaffected, because the library
creates a little-endian block for them and the value happens to read
back correctly. Every test page is such a file, which is why the suite
stayed green.

Add regression tests that use pages already carrying an EXIF block in
either byte order, follow the rotation through PageOperations.Rotate as
well, and check the result with an EXIF reader that is independent of
the library this app uses.

+++ EBookBuilder.Core/Exif/ExifOrientationStore.cs
@@ -26,10 +26,23 @@ public static class ExifOrientationStore
     /// <summary>
     /// Rewrites the Orientation of the specified file. The pixel data is not touched.
     /// </summary>
+    /// <remarks>
+    /// <para>
+    /// The value is passed as a <see cref="ushort"/> because Orientation is defined as an EXIF
+    /// SHORT. The library also has a byte overload, and passing a byte writes the tag with the
+    /// EXIF type BYTE instead: the library's own reader then reads the value in the file's byte
+    /// order, so on a big-endian (camera-style) file it came back as a value that is not a valid
+    /// orientation and every rotation looked like it had done nothing.
+    /// </para>
+    /// <para>
+    /// A BYTE-typed Orientation is also invalid as far as other software is concerned and gets
+    /// ignored, so the rotation was invisible there as well, whatever the byte order.
+    /// </para>
+    /// </remarks>
     public static void Write(string path, ExifOrientation orientation)
     {
         var file = ImageFile.FromFile(path);
-        file.Properties.Set(ExifTag.Orientation, (byte)orientation);
+        file.Properties.Set(ExifTag.Orientation, (ushort)orientation);
         file.Save(path);
     }
 }
